Update onSettingsChanged.js
I made some rearrangements to make the code easier to understand. Initially, when I looked at it, I found it a little confusing to see which settings it switched and how they were related to each other. But once I understood it by comparing it to the UI, I realized which settings do what, and I saw the reason for the confusion. So I rearranged it in a logical order to make it easier to understand from the start.